Add new version of space-based weighting algorithm
This patch introduces a new weighting algorithm that tries to combine
the strengths of both current algorithms. It is primarily space-based,
but instead of taking the total free space and apply the fragmentation
to it, it weights each bucket of segments (larger segment-size buckets
get a higher weighting, to convey that larger free segments are worth
more than just their larger size). It also includes information about
the bucket of the largest free segment, which allows us to make better
decisions about whether the metaslab is suitable for a given allocation.

Sponsored-by: Wasabi Technology, Inc.
Sponsored-by: Klara, Inc.
Signed-off-by: Paul Dagnelie <paul.dagnelie@klarasystems.com>